chore: move all OnBoarding related class to `common.onboarding`
feat: add removed method
