+++ server/src/main/kotlin/org/elaastic/common/onboarding/OnboardingState.kt
@@ -16,11 +16,10 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package org.elaastic.user
+package org.elaastic.common.onboarding
 
-import org.elaastic.common.onboarding.OnboardingChapter
-import org.elaastic.common.onboarding.OnboardingChapterConverter
 import org.elaastic.common.persistence.AbstractJpaPersistable
+import org.elaastic.user.User
 import org.springframework.data.jpa.domain.support.AuditingEntityListener
 import javax.persistence.Convert
 import javax.persistence.Entity
