+++ server/src/main/kotlin/org/elaastic/user/UserService.kt
@@ -20,6 +20,8 @@ package org.elaastic.user
 
 import org.apache.commons.lang3.time.DateUtils
 import org.elaastic.common.onboarding.OnboardingChapter
+import org.elaastic.common.onboarding.OnboardingState
+import org.elaastic.common.onboarding.OnboardingStateRepository
 import org.elaastic.user.legal.TermsService
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.security.access.AccessDeniedException
