+++ server/src/test/kotlin/org/elaastic/assignment/AssignmentControllerTest.kt
@@ -3,9 +3,9 @@ package org.elaastic.assignment
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.whenever
 import org.elaastic.common.web.MessageBuilder
-import org.elaastic.material.instructional.subject.SubjectService
 import org.elaastic.security.TestSecurityConfig
-import org.elaastic.user.OnboardingState
+import org.elaastic.material.instructional.subject.SubjectService
+import org.elaastic.common.onboarding.OnboardingState
 import org.elaastic.user.User
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
