+++ server/src/test/kotlin/org/elaastic/material/instructional/course/CourseControllerTest.kt
@@ -23,10 +23,10 @@ import com.nhaarman.mockitokotlin2.eq
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.whenever
 import org.elaastic.common.web.MessageBuilder
-import org.elaastic.material.instructional.subject.SubjectService
-import org.elaastic.security.TestSecurityConfig
-import org.elaastic.user.OnboardingState
+import org.elaastic.common.onboarding.OnboardingState
 import org.elaastic.user.User
+import org.elaastic.security.TestSecurityConfig
+import org.elaastic.material.instructional.subject.SubjectService
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
 import org.springframework.beans.factory.annotation.Autowired
