+++ server/src/test/kotlin/org/elaastic/material/instructional/subject/SubjectControllerTest.kt
@@ -8,7 +8,7 @@ import org.elaastic.material.instructional.course.CourseService
 import org.elaastic.material.instructional.question.attachment.AttachmentService
 import org.elaastic.material.instructional.statement.StatementService
 import org.elaastic.security.TestSecurityConfig
-import org.elaastic.user.OnboardingState
+import org.elaastic.common.onboarding.OnboardingState
 import org.elaastic.user.User
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
